Update user_authentication.h
Check If admin, generate an SSH key pair for the admin user with the username "admin" and move the public key to the public key directory and the private key to the key file directory. Return from the function.
Then, generate an SSH key pair for the specified username and move the public key to the public key directory and the private key to the key file directory.
Note that the function uses the system function to run the ssh-keygen command, which may not be available on all systems. Also, it assumes that the SSH key pair is generated with the RSA algorithm with a key size of 2048 bits.

Many functions in previous commits is deprecated since openssl 3.0, changed to system() for now.